Segmentation Optimizations
The segmentation class is now using the mask of the downscaled segmentation, resizing it to the original dimensions and extracting the skin directly from the original image. This way the quality/texture of the skin is completely preserved while speeding up processing using the downscaled version. The output we were originally getting from downscaling is on the left, on the right is the directly extracted skin. It's running at 15-16 FPS using my RTX 2080. After this I took off the visualization overhead "segmentation drawing function" which was unnecessarily overlaying the mask on the original image (this output is not being used at all). This bumped up the FPS up to 20-21 FPS. Also added option to select the GPU you want use in case of multiple and added a verbose option to not clutter the terminal output if not wanted. These numbers are using my webcam with a resolution of 640x480, downscaling to 256x192, 16% the original area. Using the original 640x480 resolution, it's outputting 10-11FPS. I also changed the scaling function to be mathematically accurate based on pixel area.